Asserts required environment variables are present
Piping `habash_getidoftask` result to a series of greps
prevents making the check on `hcurl` because the exit code is ignored -
and I don't know if `set -o pipefail` is POSIX, probably not.

Asserting the existence of variables before any command is executed
remains the viable option, which is why it was chosen here. Putting the
main program inside a function makes it more compliant to the rest of
the script.

Commands are not composing the call anymore because any command with
`exit 2` would make it exit with a "Invalid action" message which
doesn't tell much.
A switch|case is used to dispatch commands which makes effortless to
call commands of other (more simple) names - changes reflected on the
help. Previous command names are still supported - no BC break.

This fixes #5 on [GitHub][1]

[1]: https://github.com/nasfarley88/habash/issues/5